refactored spec generation into separate project
