+++ core/src/main/scala/vegas/DSL/EncoderDSL.scala
@@ -2,7 +2,7 @@ package vegas.DSL
 
 import monocle.Lens
 import monocle.macros.GenLens
-import vegas.spec.Spec2._
+import vegas.spec.Spec._
 import vegas.macros.{ alias_with_lens, aliased }
 
 @aliased
