+++ core/src/main/scala/vegas/package.scala
@@ -1,4 +1,4 @@
-import vegas.spec.Spec2
+import vegas.spec.Spec
 
 /**
   * Use package object to list public API
@@ -8,78 +8,78 @@ package object vegas {
   val Vegas = DSL.Vegas
   type SpecBuilder = DSL.SpecBuilder
 
-  val Json = Spec2.DataFormatTypeEnums.Json
-  val CSV = Spec2.DataFormatTypeEnums.Csv
-  val TSV = Spec2.DataFormatTypeEnums.Tsv
+  val Json = Spec.DataFormatTypeEnums.Json
+  val CSV = Spec.DataFormatTypeEnums.Csv
+  val TSV = Spec.DataFormatTypeEnums.Tsv
 
-  val Bar = Spec2.MarkEnums.Bar
-  val Circle = Spec2.MarkEnums.Circle
-  val Square = Spec2.MarkEnums.Square
-  val Tick = Spec2.MarkEnums.Tick
-  val Line = Spec2.MarkEnums.Line
-  val Area = Spec2.MarkEnums.Area
-  val Point = Spec2.MarkEnums.Point
-  val Text = Spec2.MarkEnums.Text
+  val Bar = Spec.MarkEnums.Bar
+  val Circle = Spec.MarkEnums.Circle
+  val Square = Spec.MarkEnums.Square
+  val Tick = Spec.MarkEnums.Tick
+  val Line = Spec.MarkEnums.Line
+  val Area = Spec.MarkEnums.Area
+  val Point = Spec.MarkEnums.Point
+  val Text = Spec.MarkEnums.Text
 
-  val Quantitative = Spec2.TypeEnums.Quantitative
-  val Nominal = Spec2.TypeEnums.Nominal
-  val Ordinal = Spec2.TypeEnums.Ordinal
-  val Temporal = Spec2.TypeEnums.Temporal
+  val Quantitative = Spec.TypeEnums.Quantitative
+  val Nominal = Spec.TypeEnums.Nominal
+  val Ordinal = Spec.TypeEnums.Ordinal
+  val Temporal = Spec.TypeEnums.Temporal
 
-  val Count = Spec2.AggregateOpEnums.Count
-  val Valid = Spec2.AggregateOpEnums.Valid
-  val Missing = Spec2.AggregateOpEnums.Missing
-  val Distinct = Spec2.AggregateOpEnums.Distinct
-  val Sum = Spec2.AggregateOpEnums.Sum
-  val Mean = Spec2.AggregateOpEnums.Mean
-  val Average = Spec2.AggregateOpEnums.Average
-  val Variance = Spec2.AggregateOpEnums.Variance
-  val VarianceP = Spec2.AggregateOpEnums.Variancep
-  val Stdev = Spec2.AggregateOpEnums.Stdev
-  val Median = Spec2.AggregateOpEnums.Median
-  val Q1 = Spec2.AggregateOpEnums.Q1
-  val Q3 = Spec2.AggregateOpEnums.Q3
-  val Modeskew = Spec2.AggregateOpEnums.Modeskew
-  val Min = Spec2.AggregateOpEnums.Min
-  val Max = Spec2.AggregateOpEnums.Max
+  val Count = Spec.AggregateOpEnums.Count
+  val Valid = Spec.AggregateOpEnums.Valid
+  val Missing = Spec.AggregateOpEnums.Missing
+  val Distinct = Spec.AggregateOpEnums.Distinct
+  val Sum = Spec.AggregateOpEnums.Sum
+  val Mean = Spec.AggregateOpEnums.Mean
+  val Average = Spec.AggregateOpEnums.Average
+  val Variance = Spec.AggregateOpEnums.Variance
+  val VarianceP = Spec.AggregateOpEnums.Variancep
+  val Stdev = Spec.AggregateOpEnums.Stdev
+  val Median = Spec.AggregateOpEnums.Median
+  val Q1 = Spec.AggregateOpEnums.Q1
+  val Q3 = Spec.AggregateOpEnums.Q3
+  val Modeskew = Spec.AggregateOpEnums.Modeskew
+  val Min = Spec.AggregateOpEnums.Min
+  val Max = Spec.AggregateOpEnums.Max
 
-  val Linear = Spec2.ScaleTypeEnums.Linear
-  val Log = Spec2.ScaleTypeEnums.Log
-  val Pow = Spec2.ScaleTypeEnums.Pow
-  val Sqrt = Spec2.ScaleTypeEnums.Sqrt
-  val Quantile = Spec2.ScaleTypeEnums.Quantile
-  val Quantize = Spec2.ScaleTypeEnums.Quantize
-  val Time = Spec2.ScaleTypeEnums.Time
-  val OrdinalS = Spec2.ScaleTypeEnums.Ordinal
+  val Linear = Spec.ScaleTypeEnums.Linear
+  val Log = Spec.ScaleTypeEnums.Log
+  val Pow = Spec.ScaleTypeEnums.Pow
+  val Sqrt = Spec.ScaleTypeEnums.Sqrt
+  val Quantile = Spec.ScaleTypeEnums.Quantile
+  val Quantize = Spec.ScaleTypeEnums.Quantize
+  val Time = Spec.ScaleTypeEnums.Time
+  val OrdinalS = Spec.ScaleTypeEnums.Ordinal
 
-  val Bottom = Spec2.AxisOrientEnums.Bottom
-  val Top = Spec2.AxisOrientEnums.Top
-  val Left = Spec2.AxisOrientEnums.Left
-  val Right = Spec2.AxisOrientEnums.Right
+  val Bottom = Spec.AxisOrientEnums.Bottom
+  val Top = Spec.AxisOrientEnums.Top
+  val Left = Spec.AxisOrientEnums.Left
+  val Right = Spec.AxisOrientEnums.Right
 
   val Category10 = "category10"
   val Category20 = "category20"
   val Category20b = "category20b"
   val Category20c = "category20c"
 
-  val Year = Spec2.TimeUnitEnums.Year
-  val YearMonth = Spec2.TimeUnitEnums.Yearmonth
-  val YearMonthDate = Spec2.TimeUnitEnums.Yearmonthdate
-  val YearQuarter = Spec2.TimeUnitEnums.Yearquarter
-  val YearQuarterMonth = Spec2.TimeUnitEnums.Yearquartermonth
-  val YearMonthDateHours = Spec2.TimeUnitEnums.Yearmonthdatehours
-  val YearMonthDateHoursMinutes = Spec2.TimeUnitEnums.Yearmonthdatehoursminutes
-  val YearMonthDateHoursMinutesSeconds = Spec2.TimeUnitEnums.Yearmonthdatehoursminutesseconds
-  val Month = Spec2.TimeUnitEnums.Month
-  val Day = Spec2.TimeUnitEnums.Day
-  val Date = Spec2.TimeUnitEnums.Date
-  val Hours = Spec2.TimeUnitEnums.Hours
-  val Minutes = Spec2.TimeUnitEnums.Minutes
-  val Seconds = Spec2.TimeUnitEnums.Seconds
-  val Milliseconds = Spec2.TimeUnitEnums.Milliseconds
-  val HoursMinutes = Spec2.TimeUnitEnums.Hoursminutes
-  val HoursMinutesSeconds = Spec2.TimeUnitEnums.Hoursminutesseconds
-  val MinutesSeconds = Spec2.TimeUnitEnums.Minutesseconds
-  val SecondsMilliseconds = Spec2.TimeUnitEnums.Secondsmilliseconds
+  val Year = Spec.TimeUnitEnums.Year
+  val YearMonth = Spec.TimeUnitEnums.Yearmonth
+  val YearMonthDate = Spec.TimeUnitEnums.Yearmonthdate
+  val YearQuarter = Spec.TimeUnitEnums.Yearquarter
+  val YearQuarterMonth = Spec.TimeUnitEnums.Yearquartermonth
+  val YearMonthDateHours = Spec.TimeUnitEnums.Yearmonthdatehours
+  val YearMonthDateHoursMinutes = Spec.TimeUnitEnums.Yearmonthdatehoursminutes
+  val YearMonthDateHoursMinutesSeconds = Spec.TimeUnitEnums.Yearmonthdatehoursminutesseconds
+  val Month = Spec.TimeUnitEnums.Month
+  val Day = Spec.TimeUnitEnums.Day
+  val Date = Spec.TimeUnitEnums.Date
+  val Hours = Spec.TimeUnitEnums.Hours
+  val Minutes = Spec.TimeUnitEnums.Minutes
+  val Seconds = Spec.TimeUnitEnums.Seconds
+  val Milliseconds = Spec.TimeUnitEnums.Milliseconds
+  val HoursMinutes = Spec.TimeUnitEnums.Hoursminutes
+  val HoursMinutesSeconds = Spec.TimeUnitEnums.Hoursminutesseconds
+  val MinutesSeconds = Spec.TimeUnitEnums.Minutesseconds
+  val SecondsMilliseconds = Spec.TimeUnitEnums.Secondsmilliseconds
 
 }
