+++ core/src/main/scala/vegas/render/BaseHTMLRenderer.scala
@@ -1,6 +1,6 @@
 package vegas.render
 
-import vegas.spec.Spec2.ExtendedUnitSpec
+import vegas.spec.Spec.ExtendedUnitSpec
 
 /**
   * @author Aish Fenton.
