+++ core/src/main/scala/vegas/spec/package.scala
@@ -2,7 +2,7 @@ package vegas
 
 import io.circe._
 import io.circe.syntax._
-import vegas.spec.Spec2._
+import vegas.spec.Spec._
 
 /**
   * @author Aish Fenton.
@@ -12,7 +12,7 @@ package object spec {
   val DropNullJsonPrinter = Printer.spaces2.copy(dropNullKeys = true)
 
   def toJson(spec: ExtendedUnitSpec) = {
-    import Spec2.Implicits._
+    import Spec.Implicits._
 
     spec.asJson.pretty(DropNullJsonPrinter)
   }
