+++ core/src/test/scala/vegas/DSL/DSLSpec.scala
@@ -1,13 +1,14 @@
 package vegas.DSL
 
 import monocle.macros.GenLens
+import org.scalatest.{FlatSpec, Matchers}
 import vegas._
 import vegas.BaseSpec
 
 /**
   * @author Aish Fenton.
   */
-class DSLSpec extends BaseSpec {
+class DSLSpec extends FlatSpec with Matchers {
 
   "_orElse, when composing a Lens and a Prism" should "reach modify case classes with optional values immutably" in {
     case class Person(address: Option[Address])
