+++ core/src/test/scala/vegas/fixtures/BasicPlots.scala
@@ -5,13 +5,13 @@ import vegas._
 object BasicPlots {
 
   val SimpleBarChart =
-    Vegas("A simple bar chart with embedded data").
+    Vegas("A simple bar chart with embedded data.").
     withData(Seq(
       Map("a" -> "A", "b" -> 28), Map("a" -> "B", "b" -> 55), Map("a" -> "C", "b" -> 43),
       Map("a" -> "D", "b" -> 91), Map("a" -> "E", "b" -> 81), Map("a" -> "F", "b" -> 53),
       Map("a" -> "G", "b" -> 19), Map("a" -> "H", "b" -> 87), Map("a" -> "I", "b" -> 52)
     )).
-    encodeX("a", Nominal).
+    encodeX("a", Ordinal).
     encodeY("b", Quantitative).
     mark(Bar)
 
@@ -19,16 +19,17 @@ object BasicPlots {
     Vegas("A bar chart showing the US population distribution of age groups in 2000.").
     withDataURL("https://vega.github.io/vega-editor/app/data/population.json").
     encodeY("age", Ordinal).
+    transformFilter("datum.year == 2000").
+    mark(Bar).
     scaleY(bandSize = 17).
     encodeX("people", Quantitative, aggregate=Sum).
-    axisX(title = "population").
-    mark(Bar)
+    axisX(title = "population")
 
   val GroupedBarChart =
     Vegas().
       withDataURL("https://vega.github.io/vega-editor/app/data/population.json").
       mark(Bar).
-      addTransformCalculation("gender", """ datum.sex == 2 ? "Female" : "Male" """).
+      addTransformCalculation("gender", """datum.sex == 2 ? "Female" : "Male"""").
       transformFilter("datum.year == 2000").
       encodeColumn("age", Ordinal).
         scaleColumn(padding=4.0).
@@ -41,6 +42,6 @@ object BasicPlots {
       encodeColor("gender", Nominal).
         scaleColor(rangeNominals=List("#EA98D2", "#659CCA"))
 
-  val plots = SimpleBarChart :: AggregateBarChart :: GroupedBarChart :: Nil
+  val plots: List[SpecBuilder] = SimpleBarChart :: AggregateBarChart :: GroupedBarChart :: Nil
 
 }
