+++ core/src/test/scala/vegas/spec/SpecSpec.scala
@@ -2,17 +2,17 @@ package vegas.spec
 
 import java.io.File
 
-import vegas.BaseSpec
 import io.circe._
 import io.circe.syntax._
+import org.scalatest.{FlatSpec, Matchers}
 
 import scala.io.Source
 import vegas.JsonMatchers
 
-class SpecSpec extends BaseSpec with JsonMatchers {
+class SpecSpec extends FlatSpec with Matchers with JsonMatchers {
 
-  import vegas.spec.Spec2._
-  import vegas.spec.Spec2.Implicits._
+  import vegas.spec.Spec._
+  import vegas.spec.Spec.Implicits._
 
   val examples = new File("core/src/test/resources/example-specs").listFiles.toList.filterNot(_.isDirectory)
 
@@ -34,7 +34,7 @@ class SpecSpec extends BaseSpec with JsonMatchers {
       val spec = parser.decode[VegaUnion](json)
 
       spec should be ('isRight)
-      spec.toOption.get.asJson should beNoDifferentFrom(json)
+      spec.toOption.get.asJson should beSameJsonAs(json)
     }
   }
