+++ spec/src/main/resources/vega-lite-schema.json
